Decouple the dependency on CredentialFactory from the StringResolver. Add a lot of missing JavaDoc.
