+++ core/src/main/resources/META-INF/services/nl.nn.adapterframework.util.AdditionalStringResolver
@@ -1,0 +1,1 @@
+nl.nn.adapterframework.util.CredentialResolver
