+++ core/src/test/java/nl/nn/adapterframework/util/CredentialResolverTest.java
@@ -1,0 +1,125 @@
+package nl.nn.adapterframework.util;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
+import nl.nn.adapterframework.testutil.TestFileUtils;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+
+import java.io.InputStream;
+import java.net.URL;
+import java.util.Collections;
+import java.util.Properties;
+
+class CredentialResolverTest {
+
+	Properties properties;
+
+	@BeforeEach
+	void setUp() throws Exception {
+		properties = new Properties();
+		URL propertiesURL = TestFileUtils.getTestFileURL("/StringResolver.properties");
+		assertNotNull(propertiesURL, "properties file [StringResolver.properties] not found!");
+
+		InputStream propsStream = propertiesURL.openStream();
+		properties.load(propsStream);
+		assertTrue(properties.size() > 0, "did not find any properties!");
+
+		System.setProperty("authAliases.expansion.allowed", "${allowedAliases}");
+	}
+
+	@Test
+	public void resolveUsername() {
+		// N.B. the notation ${credential:alias1/username} will work too, for some implementations of CredentialProvider, but not for all!
+		String result = StringResolver.substVars("${credential:username:alias1}", properties);
+		assertEquals("username1", result);
+
+		result = StringResolver.substVars("${credential:username:alias1}", properties, true);
+		assertEquals("${credential:username:alias1:-username1}", result);
+	}
+
+	@Test
+	public void resolvePassword1() {
+		// N.B. the notation ${credential:alias1/password} will work too, for some implementations of CredentialProvider, but not for all!
+		String result = StringResolver.substVars("${credential:password:alias1}", properties);
+		assertEquals("password1", result);
+
+		result = StringResolver.substVars("${credential:password:alias1}", properties, true);
+		assertEquals("${credential:password:alias1:-password1}", result);
+	}
+
+	@Test
+	public void resolvePassword2() {
+		String result = StringResolver.substVars("${credential:alias1}", properties); // the 'credential:' prefix defaults to return the password
+		assertEquals("password1", result);
+
+		result = StringResolver.substVars("${credential:alias1}", properties, true);
+		assertEquals("${credential:alias1:-password1}", result);
+	}
+
+	@Test
+	public void resolvePasswordOnlyAlias() {
+		String result = StringResolver.substVars("${credential:alias2}", properties);
+		assertEquals("passwordOnly", result);
+
+		result = StringResolver.substVars("${credential:alias2}", properties, true);
+		assertEquals("${credential:alias2:-passwordOnly}", result);
+	}
+
+	@Test
+	public void resolvePasswordOnlyAlias2() {
+		String result = StringResolver.substVars("${credential:password:alias2}", properties);
+		assertEquals("passwordOnly", result);
+
+		result = StringResolver.substVars("${credential:password:alias2}", properties, true);
+		assertEquals("${credential:password:alias2:-passwordOnly}", result);
+	}
+
+	@Test
+	public void resolvePasswordNotAllowed() {
+		String result = StringResolver.substVars("${credential:password:alias3}", properties);
+		assertEquals("!!not allowed to expand credential of authAlias [alias3]!!", result);
+
+		result = StringResolver.substVars("${credential:password:alias3}", properties, true);
+		assertEquals("${credential:password:alias3:-!!not allowed to expand credential of authAlias [alias3]!!}", result);
+	}
+
+	@Test
+	public void resolvePassword1AndHide() {
+		// N.B. the notation ${credential:alias1/password} will work too, for some implementations of CredentialProvider, but not for all!
+		String result = StringResolver.substVars("${credential:password:alias1}", properties, null, Collections.emptyList());
+		assertEquals("*********", result);
+
+		result = StringResolver.substVars("${credential:password:alias1}", properties, null, Collections.emptyList(), true);
+		assertEquals("${credential:password:alias1:-*********}", result);
+	}
+
+	@Test
+	public void resolvePassword2AndHide() {
+		String result = StringResolver.substVars("${credential:alias1}", properties, null, Collections.emptyList()); // the 'credential:' prefix defaults to return the password
+		assertEquals("*********", result);
+
+		result = StringResolver.substVars("${credential:alias1}", properties, null, Collections.emptyList(), true);
+		assertEquals("${credential:alias1:-*********}", result);
+	}
+
+	@Test
+	public void resolvePasswordOnlyAliasAndHide() {
+		String result = StringResolver.substVars("${credential:alias2}", properties, null, Collections.emptyList());
+		assertEquals("************", result);
+
+		result = StringResolver.substVars("${credential:alias2}", properties, null, Collections.emptyList(), true);
+		assertEquals("${credential:alias2:-************}", result);
+	}
+
+	@Test
+	public void resolvePasswordOnlyAlias2AndHide() {
+		String result = StringResolver.substVars("${credential:password:alias2}", properties, null, Collections.emptyList());
+		assertEquals("************", result);
+
+		result = StringResolver.substVars("${credential:password:alias2}", properties, null, Collections.emptyList(), true);
+		assertEquals("${credential:password:alias2:-************}", result);
+	}
+}
